# docs/api/openapi/knowledgebuilder-openapi.yaml
openapi: 3.0.3
info:
  title: KnowledgeBuilder API
  version: 1.0.0
  description: |
    KnowledgeBuilder collects and structures knowledge about entities using multi-source search 
    and fact extraction. It provides RESTful endpoints for creating, managing, and searching 
    knowledge clouds with support for deepening cycles to enhance knowledge quality.
    
    ## Key Features
    - **Entity-based knowledge generation**: Generate knowledge from entity identifiers
    - **Multi-source search**: Aggregate information from multiple sources
    - **Deepening cycles**: Iterative knowledge refinement (0-11 cycles)
    - **Semantic search**: Query knowledge using natural language
    - **Entity lookup**: Retrieve knowledge by entity identifier
    - **Full CRUD operations**: Create, read, update, and delete knowledge entries
    
    ## Architecture
    KnowledgeBuilder uses the unified API standard with standardized request/response formats,
    Bearer token authentication, comprehensive error handling, pagination, filtering, sorting,
    rate limiting, and monitoring.
    
  contact:
    name: Chrysalis API Support
    url: https://docs.chrysalis.dev
    email: support@chrysalis.dev
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5002
    description: Development server
  - url: https://api.chrysalis.dev/knowledgebuilder
    description: Production server
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/knowledge:
    post:
      tags:
        - Knowledge
      summary: Create knowledge
      description: |
        Generates knowledge cloud for a specified entity with optional deepening cycles.
        
        **Process Flow:**
        1. Validates request parameters
        2. Runs knowledge pipeline
        3. Collects facts from multiple sources
        4. Applies deepening cycles for refinement
        5. Stores knowledge with generated ID
        6. Returns complete knowledge cloud
        
        **Rate Limits:** 10 requests per minute per API key
      operationId: createKnowledge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KnowledgeCreateRequest'
            examples:
              person:
                summary: Person entity
                value:
                  identifier: Bob Ross
                  entity_type: Person
                  deepening_cycles: 3
              organization:
                summary: Organization entity
                value:
                  identifier: NASA
                  entity_type: Organization
                  deepening_cycles: 5
      responses:
        '201':
          description: Knowledge created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/APIResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Knowledge'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/ServiceError'
    
    get:
      tags:
        - Knowledge
      summary: List knowledge
      description: |
        Retrieves a paginated list of all knowledge entries with optional filtering and sorting.
        
        **Rate Limits:** 100 requests per minute per API key
      operationId: listKnowledge
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortParam'
        - name: filter[entity_type]
          in: query
          description: Filter by entity type
          schema:
            type: string
        - name: filter[identifier]
          in: query
          description: Filter by identifier
          schema:
            type: string
      responses:
        '200':
          description: List of knowledge entries retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/APIResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Knowledge'
                      meta:
                        allOf:
                          - $ref: '#/components/schemas/ResponseMeta'
                          - type: object
                            properties:
                              pagination:
                                $ref: '#/components/schemas/PaginationMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServiceError'
  /api/v1/knowledge/{knowledge_id}:
    get:
      tags:
        - Knowledge
      summary: Get knowledge by ID
      description: Retrieves a specific knowledge entry by its unique identifier
      operationId: getKnowledge
      parameters:
        - $ref: '#/components/parameters/KnowledgeIdParam'
      responses:
        '200':
          description: Knowledge entry retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/APIResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Knowledge'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServiceError'
    
    patch:
      tags:
        - Knowledge
      summary: Partially update knowledge
      description: |
        Updates specific fields of an existing knowledge entry without replacing the entire resource.
        Only provided fields will be updated.
      operationId: updateKnowledge
      parameters:
        - $ref: '#/components/parameters/KnowledgeIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KnowledgeUpdateRequest'
            examples:
              update_identifier:
                summary: Update identifier
                value:
                  identifier: Robert Norman Ross
              update_cycles:
                summary: Update deepening cycles
                value:
                  deepening_cycles: 5
      responses:
        '200':
          description: Knowledge entry updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/APIResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Knowledge'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/ServiceError'
    
    put:
      tags:
        - Knowledge
      summary: Replace knowledge
      description: |
        Completely replaces an existing knowledge entry with new data.
        All required fields must be provided.
      operationId: replaceKnowledge
      parameters:
        - $ref: '#/components/parameters/KnowledgeIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KnowledgeReplaceRequest'
      responses:
        '200':
          description: Knowledge entry replaced successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/APIResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Knowledge'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/ServiceError'
    
    delete:
      tags:
        - Knowledge
      summary: Delete knowledge
      description: Permanently deletes a knowledge entry
      operationId: deleteKnowledge
      parameters:
        - $ref: '#/components/parameters/KnowledgeIdParam'
      responses:
        '200':
          description: Knowledge entry deleted successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/APIResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          deleted:
                            type: boolean
                            example: true
                          knowledge_id:
                            type: string
                            example: person:bob_ross
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServiceError'
  /api/v1/knowledge/search:
    post:
      tags:
        - Search
      summary: Search knowledge
      description: |
        Advanced semantic search for knowledge entries using natural language queries.
        
        **Rate Limits:** 30 requests per minute per API key
      operationId: searchKnowledge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KnowledgeSearchRequest'
            examples:
              basic_search:
                summary: Basic search
                value:
                  query: American painters from the 20th century
                  limit: 10
              filtered_search:
                summary: Filtered search
                value:
                  query: space exploration
                  entity_type: Organization
                  limit: 5
      responses:
        '200':
          description: Search results retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/APIResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          query:
                            type: string
                            example: American painters from the 20th century
                          results:
                            type: array
                            items:
                              allOf:
                                - $ref: '#/components/schemas/Knowledge'
                                - type: object
                                  properties:
                                    relevance_score:
                                      type: number
                                      format: float
                                      minimum: 0
                                      maximum: 1
                                      example: 0.94
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/ServiceError'
  /api/v1/knowledge/entities/{entity_id}:
    get:
      tags:
        - Knowledge
      summary: Get knowledge by entity
      description: |
        Retrieves knowledge entries for a specific entity identifier.
        
        **Note:** If multiple knowledge entries exist for the same entity, returns an array.
      operationId: getKnowledgeByEntity
      parameters:
        - name: entity_id
          in: path
          required: true
          description: Entity identifier (case-insensitive)
          schema:
            type: string
            example: bob_ross
      responses:
        '200':
          description: Knowledge entry retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/APIResponse'
                  - type: object
                    properties:
                      data:
                        oneOf:
                          - $ref: '#/components/schemas/Knowledge'
                          - type: array
                            items:
                              $ref: '#/components/schemas/Knowledge'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServiceError'
  /knowledge:
    post:
      tags:
        - Knowledge
      summary: Legacy create endpoint (deprecated)
      description: |
        **⚠️ DEPRECATED**: Use POST /api/v1/knowledge instead.
        
        This endpoint will be removed in v2.0.
        
        **Migration Guide:**
        1. Remove `apiKeys` from request body
        2. Add `Authorization: Bearer <api_key>` header
        3. Use `/api/v1/knowledge` endpoint
        4. Handle standardized response format
      operationId: legacyCreateKnowledge
      deprecated: true
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - identifier
              properties:
                identifier:
                  type: string
                entity_type:
                  type: string
                deepening_cycles:
                  type: integer
                  minimum: 0
                apiKeys:
                  type: object
                  properties:
                    OPENAI_API_KEY:
                      type: string
                    ANTHROPIC_API_KEY:
                      type: string
      responses:
        '201':
          description: Knowledge created (with deprecation warning)
          content:
            application/json:
              schema:
                type: object
                properties:
                  knowledge_items:
                    type: array
                    items:
                      type: object
                  meta:
                    type: object
                    properties:
                      deprecated:
                        type: boolean
                        example: true
                      deprecation_notice:
                        type: string
                        example: Use POST /api/v1/knowledge instead
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Bearer token authentication using API keys or JWT tokens.
        
        **Format:** `Authorization: Bearer <token>`
        
        **API Key Format:** `<keyId>.<secret>`
  parameters:
    KnowledgeIdParam:
      name: knowledge_id
      in: path
      required: true
      description: Unique knowledge identifier (format entity_type:identifier)
      schema:
        type: string
        example: person:bob_ross
    PageParam:
      name: page
      in: query
      description: Page number for pagination (1-indexed)
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      description: Number of items per page (max 100)
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortParam:
      name: sort
      in: query
      description: |
        Comma-separated list of fields to sort by.
        Prefix with `-` for descending order.
        
        **Examples:**
        - `identifier` - Sort by identifier ascending
        - `-deepening_cycles` - Sort by cycles descending
        - `entity_type,-identifier` - Multi-field sort
      schema:
        type: string
        example: -deepening_cycles
  schemas:
...
    KnowledgeCreateRequest:
      type: object
      required:
        - identifier
      properties:
        identifier:
          type: string
          minLength: 1
          description: Entity identifier (name, ID, etc.)
          example: Bob Ross
        entity_type:
          type: string
          description: Type of entity (Person, Organization, etc.)
          example: Person
        deepening_cycles:
          type: integer
          minimum: 0
          default: 0
          description: Number of iterative refinement cycles
          example: 3
...
    APIResponse:
      type: object
      required:
        - success
        - meta
      properties:
        success:
          type: boolean
          description: Indicates if the request was successful
        data:
          description: Response data (present on success)
        error:
          $ref: '#/components/schemas/APIError'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    Knowledge:
      type: object
      properties:
        knowledge_id:
          type: string
          description: Generated from entity_type:identifier
          example: person:bob_ross
        identifier:
          type: string
          example: Bob Ross
        entity_type:
          type: string
          example: Person
        knowledge_items:
          type: array
          items:
            $ref: '#/components/schemas/KnowledgeItem'
        deepening_cycles:
          type: integer
          minimum: 0
          example: 3
        created_at:
          type: string
          format: date-time
          example: '2026-01-11T04:00:00.000Z'
        updated_at:
          type: string
          format: date-time
          example: '2026-01-11T05:00:00.000Z'
...
    ResponseMeta:
      type: object
      required:
        - request_id
        - timestamp
        - version
      properties:
        request_id:
          type: string
          format: uuid
          description: Unique request identifier
        timestamp:
          type: string
          format: date-time
          description: Response timestamp
        version:
          type: string
          description: API version
          example: v1
        deprecated:
          type: boolean
          description: Indicates if endpoint is deprecated
        deprecation_notice:
          type: string
          description: Deprecation message
    PaginationMeta:
      type: object
      required:
        - page
        - per_page
        - total
        - total_pages
        - has_next
        - has_prev
      properties:
        page:
          type: integer
          minimum: 1
          description: Current page number (1-indexed)
        per_page:
          type: integer
          minimum: 1
          maximum: 100
          description: Items per page
        total:
          type: integer
          minimum: 0
          description: Total number of items
        total_pages:
          type: integer
          minimum: 0
          description: Total number of pages
        has_next:
          type: boolean
          description: Whether there is a next page
        has_prev:
          type: boolean
          description: Whether there is a previous page
  responses:
    Unauthorized:
      description: Unauthorized - Missing or invalid authentication
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/APIResponse'
              - type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        example: AUTHENTICATION_ERROR.MISSING_AUTHORIZATION
                      message:
                        type: string
                        example: Missing Authorization header
                      category:
                        type: string
                        example: AUTHENTICATION_ERROR
...
    ValidationError:
      description: Validation error - Invalid request data
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/APIResponse'
              - type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        example: VALIDATION_ERROR.REQUIRED_FIELD
                      message:
                        type: string
                        example: Field 'identifier' is required
                      category:
                        type: string
                        example: VALIDATION_ERROR
                      details:
                        type: array
                        items:
                          type: object
                          properties:
                            field:
                              type: string
                              example: identifier
                            code:
                              type: string
                              example: REQUIRED_FIELD
                            message:
                              type: string
                              example: This field is required
    ServiceError:
      description: Internal server error
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/APIResponse'
              - type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        example: SERVICE_ERROR.PIPELINE_ERROR
                      message:
                        type: string
                        example: Knowledge pipeline failure
                      category:
                        type: string
                        example: SERVICE_ERROR